split courses page into server and client parts for SEO
